Refactored cache implementations
