+++ src/base/controller/info.ts
@@ -1,4 +1,4 @@
 export type CollectionInfo = {
   dbName: string,
   collectionName: string,
-}
+};
