+++ src/utils/cache/const.ts
@@ -1,0 +1,1 @@
+export const CACHE_LIFE_SECS = 300; // 5 mins
