+++ src/utils/cache/types.ts
@@ -1,4 +1,4 @@
 export type ResourceCache<T> = {
   lastFetchedEpoch: number,
   data: T,
-}
+};
