--- src/utils/resources/const.ts
@@ -1,1 +1,0 @@
-export const CACHE_LIFE_SECS = 600; // 10 min
