+++ src/utils/resources/loader/cache/main.ts
@@ -2,12 +2,9 @@ import fetch from 'node-fetch';
 
 import {CharaInfo, DragonInfo, ResourcePaths, toUnitInfoMap} from '../../../../api-def/resources';
 import {UnitNameRefController} from '../../../../endpoints/data/unitNameRef/controller';
-import {CACHE_LIFE_SECS} from '../../const';
-import {ResourceCache} from '../../types';
+import {generateCacheBody, isCacheExpired} from '../../../cache/func';
 import {Cache, CacheKey, CacheRequireClientType} from './types';
 
-
-const generateCacheBody = <T>(lastFetchedEpoch: number, data: T): ResourceCache<T> => ({lastFetchedEpoch, data});
 
 const generateNewCache = (): Cache => ({
   [CacheKey.UNIT_INFO]: generateCacheBody(0, new Map()),
@@ -16,16 +13,12 @@ const generateNewCache = (): Cache => ({
 
 let cache: Cache = generateNewCache();
 
-const isCacheExpired = (cacheKey: CacheKey, currentEpoch: number) => {
-  return currentEpoch - cache[cacheKey].lastFetchedEpoch > CACHE_LIFE_SECS;
-};
-
 const setCache = async <K extends CacheKey>(
   epoch: number,
   mongoClient: CacheRequireClientType<K>,
 ): Promise<void> => {
   // Fetch unit info data if not yet fetched
-  if (!cache[CacheKey.UNIT_INFO].data.size || isCacheExpired(CacheKey.UNIT_INFO, epoch)) {
+  if (!cache[CacheKey.UNIT_INFO].data.size || isCacheExpired(cache[CacheKey.UNIT_INFO], epoch)) {
     const [charaInfo, dragonInfo] = await Promise.all([
       fetch(ResourcePaths.INFO_CHARA)
         .then((response) => response.json()) as unknown as CharaInfo,
@@ -60,7 +53,7 @@ export const getCache = async <K extends CacheKey>(
 ): Promise<Cache[K]['data']> => {
   const currentEpoch = Math.round(Date.now() / 1000);
 
-  if (isCacheExpired(cacheKey, currentEpoch)) {
+  if (isCacheExpired(cache[cacheKey], currentEpoch)) {
     await setCache<K>(currentEpoch, mongoClient);
   }
