+++ src/utils/resources/loader/cache/types.ts
@@ -2,7 +2,7 @@ import {MongoClient} from 'mongodb';
 
 import {UnitNameRefData} from '../../../../api-def/api';
 import {UnitInfoMap} from '../../../../api-def/resources';
-import {ResourceCache} from '../../types';
+import {ResourceCache} from '../../../cache/types';
 
 
 export enum CacheKey {
@@ -13,11 +13,11 @@ export enum CacheKey {
 export type Cache = {
   [CacheKey.UNIT_INFO]: ResourceCache<UnitInfoMap<number>>,
   [CacheKey.UNIT_NAME_2_ID]: ResourceCache<UnitNameRefData>,
-}
+};
 
 type CacheRequireClient = {
   [CacheKey.UNIT_INFO]: false,
   [CacheKey.UNIT_NAME_2_ID]: true,
-}
+};
 
 export type CacheRequireClientType<K extends CacheKey> = CacheRequireClient[K] extends true ? MongoClient : undefined;
